rework keyboard handling VC
